+++ include/onmt/Threads.h
@@ -1,0 +1,13 @@
+#pragma once
+
+namespace onmt
+{
+
+  class Threads
+  {
+  public:
+    static void set(int number);
+    static int get();
+  };
+
+}