+++ src/Threads.cc
@@ -1,0 +1,18 @@
+#include "onmt/Threads.h"
+
+#include <Eigen/Core>
+
+namespace onmt
+{
+
+  void Threads::set(int number)
+  {
+    return Eigen::setNbThreads(number);
+  }
+
+  int Threads::get()
+  {
+    return Eigen::nbThreads();
+  }
+
+}