+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## [Unreleased]
+
+### New features
+
+* Programmatically set the number of threads to use
+
 ## [v0.1.0](https://github.com/OpenNMT/CTranslate/releases/tag/v0.1.0) (2017-01-11)
 
 Initial release.